ENG-3788 Refactoring
