+++ src/commands/api/add.ts
@@ -4,10 +4,11 @@ import { ApiClaimService } from '../../services/api-claim-service'
 import { BundleService } from '../../services/bundle-service'
 
 export default class Add extends Command {
-  static description = 'Adds an internal API claim to the specified MFE component'
+  static description =
+    'Adds an internal API claim to the specified MFE component'
 
   static examples = [
-    '<%= config.bin %> <%= command.id %> mfe1 ms1-api --serviceId ms1 --serviceUrl http://localhost:8080',
+    '<%= config.bin %> <%= command.id %> mfe1 ms1-api --serviceId ms1 --serviceUrl http://localhost:8080'
   ]
 
   static args = [
@@ -35,7 +36,7 @@ export default class Add extends Command {
   }
 
   public async run(): Promise<void> {
-    BundleService.verifyBundleInitialized(process.cwd())
+    BundleService.isValidBundleProject(process.cwd())
 
     const { args, flags } = await this.parse(Add)
 
@@ -46,9 +47,14 @@ export default class Add extends Command {
     }
     const apiClaimService: ApiClaimService = new ApiClaimService()
 
-    CliUx.ux.action.start(`Adding a new API claim named ${args.claimName} to Micro Frontend ${args.mfeName}`)
-    apiClaimService.addInternalApiClaim(args.mfeName, apiClaim, flags.serviceUrl)
+    CliUx.ux.action.start(
+      `Adding a new API claim named ${args.claimName} to Micro Frontend ${args.mfeName}`
+    )
+    apiClaimService.addInternalApiClaim(
+      args.mfeName,
+      apiClaim,
+      flags.serviceUrl
+    )
     CliUx.ux.action.stop()
   }
 }
-
