+++ src/commands/build.ts
@@ -24,7 +24,7 @@ export default class Build extends BaseBuildCommand {
   }
 
   public async run(): Promise<void> {
-    BundleService.verifyBundleInitialized(process.cwd())
+    BundleService.isValidBundleProject(process.cwd())
     const { args, flags } = await this.parse(Build)
 
     const componentService = new ComponentService()
