+++ src/commands/list.ts
@@ -22,7 +22,7 @@ export default class List extends Command {
   }
 
   public async run(): Promise<void> {
-    BundleService.verifyBundleInitialized(process.cwd())
+    BundleService.isValidBundleProject(process.cwd())
 
     const { flags } = await this.parse(List)
     const componentService = new ComponentService()
