+++ src/commands/mfe/rm.ts
@@ -16,7 +16,7 @@ export default class Rm extends Command {
   ]
 
   public async run(): Promise<void> {
-    BundleService.verifyBundleInitialized(process.cwd())
+    BundleService.isValidBundleProject(process.cwd())
 
     const { args } = await this.parse(Rm)
