+++ src/commands/svc/disable.ts
@@ -18,7 +18,7 @@ export default class Disable extends Command {
   static flags = {}
 
   public async run(): Promise<void> {
-    BundleService.verifyBundleInitialized(process.cwd())
+    BundleService.isValidBundleProject(process.cwd())
 
     const { args } = await this.parse(Disable)
