+++ src/commands/svc/enable.ts
@@ -18,7 +18,7 @@ export default class Enable extends Command {
   static flags = {}
 
   public async run(): Promise<void> {
-    BundleService.verifyBundleInitialized(process.cwd())
+    BundleService.isValidBundleProject(process.cwd())
 
     const { args } = await this.parse(Enable)
