+++ src/commands/svc/list.ts
@@ -15,7 +15,7 @@ export default class List extends Command {
   }
 
   public async run(): Promise<void> {
-    BundleService.verifyBundleInitialized(process.cwd())
+    BundleService.isValidBundleProject(process.cwd())
 
     const { flags } = await this.parse(List)
