+++ test/commands/api/add.test.ts
@@ -8,11 +8,14 @@ import {
   MicroFrontend,
   Microservice
 } from '../../../src/models/bundle-descriptor'
-import { BundleDescriptorService } from '../../../src/services/bundle-descriptor-service'
+import {
+  BundleDescriptorService,
+  MISSING_DESCRIPTOR_ERROR
+} from '../../../src/services/bundle-descriptor-service'
 import { MfeConfigService } from '../../../src/services/mfe-config-service'
 import { MfeConfig } from '../../../src/models/mfe-config'
 import { TempDirHelper } from '../../helpers/temp-dir-helper'
-import { ComponentHelper } from '../../helpers/mocks/components'
+import { ComponentHelper } from '../../helpers/mocks/component-helper'
 
 describe('api add', () => {
   const tempDirHelper = new TempDirHelper(__filename)
@@ -267,7 +270,7 @@ describe('api add', () => {
       'http://localhost:8080'
     ])
     .catch(error => {
-      expect(error.message).to.contain('not an initialized Bundle project')
+      expect(error.message).to.contain(MISSING_DESCRIPTOR_ERROR)
     })
     .it('exits with an error if current folder is not a Bundle project')
 })
