+++ test/commands/list.test.ts
@@ -7,7 +7,10 @@ import {
   MicroFrontend,
   Microservice
 } from '../../src/models/bundle-descriptor'
-import { BundleDescriptorService } from '../../src/services/bundle-descriptor-service'
+import {
+  BundleDescriptorService,
+  MISSING_DESCRIPTOR_ERROR
+} from '../../src/services/bundle-descriptor-service'
 import { TempDirHelper } from '../helpers/temp-dir-helper'
 import * as sinon from 'sinon'
 import { ConstraintsValidatorService } from '../../src/services/constraints-validator-service'
@@ -151,7 +154,7 @@ describe('list', () => {
     })
     .command(['list'])
     .catch(error => {
-      expect(error.message).to.contain('not an initialized Bundle project')
+      expect(error.message).to.contain(MISSING_DESCRIPTOR_ERROR)
     })
     .it('exits with an error if current folder is not a Bundle project')
 })
