+++ test/commands/ms/add.test.ts
@@ -7,9 +7,12 @@ import {
   Microservice,
   MicroFrontend
 } from '../../../src/models/bundle-descriptor'
-import { BundleDescriptorService } from '../../../src/services/bundle-descriptor-service'
+import {
+  BundleDescriptorService,
+  MISSING_DESCRIPTOR_ERROR
+} from '../../../src/services/bundle-descriptor-service'
 import { TempDirHelper } from '../../helpers/temp-dir-helper'
-import { ComponentHelper } from '../../helpers/mocks/components'
+import { ComponentHelper } from '../../helpers/mocks/component-helper'
 
 describe('ms add', () => {
   const bundleDescriptor: BundleDescriptor = {
@@ -144,7 +147,7 @@ describe('ms add', () => {
     })
     .command(['ms add', 'ms-in-notbundleproject'])
     .catch(error => {
-      expect(error.message).to.contain('not an initialized Bundle project')
+      expect(error.message).to.contain(MISSING_DESCRIPTOR_ERROR)
     })
     .it('exits with an error if current folder is not a Bundle project')
