+++ test/commands/ms/rm.test.ts
@@ -2,7 +2,10 @@ import { expect, test } from '@oclif/test'
 import * as fs from 'node:fs'
 import * as path from 'node:path'
 import { TempDirHelper } from '../../helpers/temp-dir-helper'
-import { BundleDescriptorService } from '../../../src/services/bundle-descriptor-service'
+import {
+  BundleDescriptorService,
+  MISSING_DESCRIPTOR_ERROR
+} from '../../../src/services/bundle-descriptor-service'
 import { MicroserviceService } from '../../../src/services/microservice-service'
 import { Microservice } from '../../../src/models/bundle-descriptor'
 import { MicroserviceStack } from '../../../src/models/component'
@@ -59,7 +62,7 @@ describe('Remove Microservice', () => {
     .do(() => tempDirHelper.createUninitializedBundleDir())
     .command(['ms rm', 'test-bundle'])
     .catch(error => {
-      expect(error.message).to.contain('is not an initialized Bundle project')
+      expect(error.message).to.contain(MISSING_DESCRIPTOR_ERROR)
     })
     .it('Returns error if Bundle directory not initialized')
 })
